refactor: move all dissipation tests to one file
